API reference

This page documents all modules and functions in BurgersJoint.

BurgersEquation

BurgersJoint.BurgersEquation.analytical_slope — Method

analytical_slope(x, t; nu=viscosity(), n=50)

Computes the analytical slope from the analytical solution using central differences.

BurgersJoint.BurgersEquation.analytical_solution — Method

analytical_solution(x, t; nu=viscosity(), n=50)

Computes the analytical solution $u(x,t)$` of Burgers' equation using the convolution form from Cole's transformation and Hermite integration. This assumes initial condition u(x,0) = -sin(pi*x).

The solution is expressed as

$u(x,t) = \frac{I_1(x,t)}{I_2(x,t)}$,

where

$I_1(x,t) = -\int_{-\infty}^{+\infty} \sin(\pi(x-\eta))\, f(x-\eta) \exp\left(-\frac{\eta^2}{4 \nu t}\right)\, \mathrm{d}\eta$,

$I_2(x,t) = \int_{-\infty}^{+\infty} f(x-\eta) \exp\left(-\frac{\eta^2}{4 \nu t}\right)\, \mathrm{d}\eta$.

Here, $f(y) = \exp(-\cos(\pi y)/(2\pi\nu))$. A change of variables $\eta = S z$ with $S = \sqrt{4 \nu t}$ is used so that

$d\eta = S \,\mathrm{d}z, \quad \exp\left(-\frac{\eta^2}{4\nu t}\right) = \exp(-z^2)$.

Hermite quadrature (with $n$ nodes, default $n = 50$) is used to evaluate these integrals.

BurgersJoint.BurgersEquation.burgers_ic — Method

burgers_ic(x)

Default initial condition for Burgers’ equation, matching the setup in Basdevant1986:

``u(x, 0) = -\sin(\pi x)``.

BurgersJoint.BurgersEquation.f — Method

f(y; nu=viscosity())

Define the function $f(y)$ as given in Basdevant1986.

BurgersJoint.BurgersEquation.viscosity — Method

viscosity()

Returns the kinematic viscosity used in Basdevant1986: $\nu = 1/(100\pi)$.

ChebyshevTau

BurgersJoint.ChebyshevTau.solveChebyshevTau — Method

solveChebyshevTau(N; dt=1e-3, T=1.0, nu=viscosity(), ic=burgers_ic)

Solves the 1D Burgers equation on t in [0, T], x in [-1, 1] using the Chebyshev-Tau spectral method with the ABCN scheme: Adams-Bashforth for the advective term and Crank-Nicholson for the viscous term, as described in Basdevant et al. (1986). As Basdevant1986 omits many details, I have also consulted the notes on spectral methods by Prof. Herman Clerx. Chebyshev-Gauss-Lobatto collocation points are used for the spatial grid.

Arguments:

N: Number of grid points - 1 (I don't like this notation, but it's common in the literature).

Keyword arguments:

dt: Time step.

T: Duration to solve over.

dt_snapshot: Time step for storing snapshots (default = -1, only initial state).

nu: Kinematic viscosity (defaults to BurgersEquation.viscosity()).

ic: Initial condition (defaults to BurgersEquation.burgers_ic()).

Output:

x: Real-space grid of length N.

u_final: Final solution in real space (length N).

snapshot_times: Vector of time steps at which the solution was saved.

snapshots: Vector of solution snapshots at times t = {n*dt_snapshot}, n=0,1,2,...

FiniteDifference

BurgersJoint.FiniteDifference.solveFiniteDifference — Method

solveFiniteDifferences(N; dt=1e-3, T=1.0, nu=viscosity(), ic=burgers_ic, S=4.0)

Solves the 1D Burgers equation ut + u ux = nu u_xx on the domain x in [-1,1] with u(+-1,t)=0 using a finite-difference method on a nonuniform grid.

The grid is generated by the coordinate transformation x = g(zeta) = { -1 + tanh(2S zeta)/tanh(S), for zeta in [0,0.5[ 1 + tanh(2S (zeta-1))/tanh(S), for zeta in [0.5,1]}, with stretching parameter S. This clusters points near the large gradient region around x = 0.

The time-stepping is done using three-point backward scheme (BDF2 with bootstrapping using BDF1).

Arguments:

N: Number of grid points

Keyword arguments:

dt: Time step.

T: Duration to solve over.

dt_snapshot: Time step for storing snapshots (default = -1, only initial state).

nu: Kinematic viscosity (defaults to BurgersEquation.viscosity()).

ic: Initial condition (defaults to BurgersEquation.burgers_ic()).

S: Stretching parameter for the grid (default = 4.0).

Output:

x: Real-space grid of length N.

u_final: Final solution in real space (length N).

snapshot_times: Vector of time steps at which the solution was saved.

snapshots: Vector of solution snapshots at times t = {n*dt_snapshot}, n=0,1,2,...

FourierSpectral

BurgersJoint.FourierSpectral.solveFourierSpectral — Method

solveFourierSpectral(N; dt=1e-3, T=1.0, nu=viscosity(), ic=burgers_ic, dealias_factor=2/3)

Solves the 1D Burgers equation on t in [0, T], x in [-1, 1] using Fourier methods (Galerkin or pseudospectral). It is based on the method described in Basdevant et al. (1986) and uses their leapfrog-like scheme for time integration.

Arguments:

N: Number of grid points (also number of Fourier modes before dealiasing).

Keyword arguments:

dt: Time step.

T: Duration to solve over.

dt_snapshot: Time step for storing snapshots (default = -1, only initial state).

nu: Kinematic viscosity (defaults to BurgersEquation.viscosity()).

ic: Initial condition (defaults to BurgersEquation.burgers_ic()).

dealias_factor: Factor for dealiasing

2/3 corresponsd to 2/3-rule Fourier Galerkin (default, aliasing errors removed).

1 corresponds to Fourier pseudospectral (aliasing errors will be present).

Output:

x: Real-space grid of length N.

u_final: Final solution in real space (length N).

snapshot_times: Vector of time steps at which the solution was saved.

snapshots: Vector of solution snapshots at times t = {j*dt_snapshot}, j=0,1,2,...

NeuralNetwork

BurgersJoint.NeuralNetwork.solveNeuralNetwork — Method

solveNeuralNetwork(N::Int; dt=1e-2, T=1.0, nu=viscosity(), ic=burgers_ic)

Solves the 1D Burgers equation on t in [0, T], x in [-1, 1] using a physics-informed neural network (PINN). If no trained model exists, it will call trainNeuralNetwork(N; dt=dt, T=T, nu=nu, ic=ic) first. The predicted solution is evaluated at a grid of N points in x and at times t = 0, 1/pi, 2/pi, ..., T.

Arguments:

N: Number of grid points in the spatial domain.

Keyword arguments (optional):

dt: Time step for the training (default = 1e-2).

T: Duration to solve over (default = 1.0).

nu: Kinematic viscosity, only used for training (default = BurgersEquation.viscosity()).

ic: Initial condition, only used for training (default = BurgersEquation.burgers_ic()).

Output:

x: Spatial grid of length N.

u_final: Final solution in real space (length N).

snapshot_times: Vector of time steps at which the solution was evaluated.

history: Vector of solution snapshots at times t = n/pi, n=0,1,2,...

BurgersJoint.NeuralNetwork.trainNeuralNetwork — Method

trainNeuralNetwork(N::Int; dt=1e-2, T=1.0, nu=viscosity(), ic=burgers_ic)

Train a PINN for the 1D Burgers equation on t in [0, T], x in [-1, 1]. Discretizes the domain with a time-step dt and a spatial-step dx = 2/(N-1). Stores the trained model in the global variable trainedModel.